Added Twillo modifications made by Arif. Modified controller logic so notification will only be sent once for each time unlocked/locked/unlocked sequence occurs. Modified controller logic so 'room led' only turns on if there is limited light detected in the room. Updated arduino device 1 to use RGB LED to better simulate airconditioner state, red for heating, blue for cooling and green for off.

+++ Arduino/device2/device2.ino
@@ -40,9 +40,9 @@ void setup() {
 
 // the loop routine runs over and over again forever:
 void loop() {
-  int totalSensorValues = 0; //store total of values read from photoresistor - used to calculate average
-  int totalReadings = 100; //total readings to take from photoresistor before calculating average
-  int avgSensorValue; //store the average sensor value and use that as our 'reading' of light
+  float totalSensorValues = 0; //store total of values read from photoresistor - used to calculate average
+  float totalReadings = 100; //total readings to take from photoresistor before calculating average
+  float avgSensorValue; //store the average sensor value and use that as our 'reading' of light
   for (int i = 0; i < totalReadings; i++) { //add totalReadings number of values from photoresistor with 1ms delay in between each reading
     totalSensorValues += analogRead(PHOTOPIN);
     delay(1);  // delay in between reads for stability
